Added a few more things..
It still doesn't function fully, I have it tar'ing, and we can almost start to use it.

I need to fix up the ignoring, currently I can have it ignore one name.. need more.

Lastly, This include is a BLOODY MESS. Structure wise, we need to discuss this.

Cheers, Narcle, we are close ;) Write up the .sex.


git-svn-id: http://simbareflection.googlecode.com/svn/trunk@21 e8c6cf14-2432-c3df-baee-86675e33b74e