Abyss - set angle North after exiting.
Prison Pete - when the solver fails (which will be 99% of the time; it's hard to solve due to the quick expire of the random), script will teleport home if SRL's reincarnate=True. (Note: if you use Reincarnate:=True and your script does not have deathwalk, you are doing it wrong!)

Reflection main file - removed old GoogleCode repo, made more user-friendly, compressed the functions a bit. To-do: use official git when Simba supports SSL.